support send tx to orderer

+++ handler/handler.go
@@ -4,16 +4,16 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/tylerztl/fabric-mempool/conf"
-	"github.com/tylerztl/fabric-mempool/protoutil"
 	"io/ioutil"
 	"math/big"
 	"os"
 
 	pb "github.com/hyperledger/fabric/protos/common"
-	//"github.com/tendermint/tendermint/types"
 	"github.com/tendermint/tendermint/config"
+	"github.com/tendermint/tendermint/types"
+	"github.com/tylerztl/fabric-mempool/conf"
 	"github.com/tylerztl/fabric-mempool/mempool"
+	"github.com/tylerztl/fabric-mempool/protoutil"
 )
 
 type Handler struct {
@@ -67,66 +67,70 @@ func (h *Handler) FetchTransactions(ctx context.Context, ftx *pb.FetchTxsRequest
 		return &pb.FetchTxsResponse{TxNum: 0, IsEmpty: true}, nil
 	}
 
-	expectedTxs := int(ftx.TxNum)
+	// TODO 分配orderer交易数量
+	expectedTxs := 10
 
 	txs := h.Mempool.ReapMaxTxs(expectedTxs - 1)
 	actualTxs := len(txs)
 	isEmpty := actualTxs < expectedTxs
 
-	logger.Info("orderer fetch transactions", "orderer", ftx.Sender,
+	logger.Info("orderer fetch transactions", "OrdererName", ftx.Requester,
 		"actualTxs", actualTxs, "expectedTxs", expectedTxs, "mempool", h.Mempool.Size())
 
-	orderer := h.fetcher.GetOrderer(ftx.Sender)
+	orderer := h.fetcher.GetOrderer(ftx.Requester)
 	if orderer == nil {
 		return nil, errors.New("not found orderer connected client")
 	}
 
-	for _, tx := range txs {
-		fee, err := protoutil.GetTxFeeFromEnvelope(tx)
+	for i, tx := range txs {
+		fee, txId, err := protoutil.GetTxFeeFromEnvelope(tx)
 		if err != nil {
 			fmt.Printf("Unmarshal tx failed: %s", err)
 			continue
 		}
 		h.distribute(fee, orderer)
-		logger.Info("Unmarshal tx", "fee", fee)
+		logger.Info("orderer fetch txs", "total", len(txs), "index", i, "txId", txId, "fee", fee)
 	}
 
-	// TODO
-	//go func() {
-	//	committedTxs := make(types.Txs, 0)
-	//	for _, tx := range txs {
-	//		err := orderer.broadcast(tx)
-	//		if err != nil {
-	//			logger.Error("failed to broadcast endorsed tx to orderer service", "error", err)
-	//			if err = orderer.resetConnect(); err == nil && orderer.broadcast(tx) != nil {
-	//				logger.Error("retry broadcast endorsed tx to orderer service", "orderer", ftx.Sender)
-	//				continue
-	//			}
-	//		}
-	//
-	//		committedTxs = append(committedTxs, tx)
-	//	}
-	//	if err := h.Mempool.Update(1, committedTxs, nil, nil, nil); err != nil {
-	//		logger.Error("txs committed update failed", "error", err)
-	//	}
-	//}()
+	go func() {
+		committedTxs := make(types.Txs, 0)
+		for _, tx := range txs {
+			err := orderer.broadcast(tx)
+			if err != nil {
+				logger.Error("failed to broadcast endorsed tx to orderer service", "error", err)
+				if err = orderer.resetConnect(); err == nil && orderer.broadcast(tx) != nil {
+					logger.Error("retry broadcast endorsed tx to orderer service", "ordererName", ftx.Requester)
+					continue
+				}
+			}
 
-	if err := h.Mempool.Update(1, txs, nil, nil, nil); err != nil {
-		logger.Error("txs committed update failed", "error", err)
-	}
+			committedTxs = append(committedTxs, tx)
+		}
+		if err := h.Mempool.Update(int64(ftx.BlockHeight), committedTxs, nil, nil, nil); err != nil {
+			logger.Error("txs committed update failed", "error", err)
+		}
+	}()
+
+	//if err := h.Mempool.Update(1, txs, nil, nil, nil); err != nil {
+	//	logger.Error("txs committed update failed", "error", err)
+	//}
 
 	return &pb.FetchTxsResponse{TxNum: int32(actualTxs), IsEmpty: isEmpty}, nil
 }
 
 func NewHandler(distributeConfig *conf.DistributeConfig) *Handler {
 	// create a unique, concurrency-safe test directory under os.TempDir()
-	rootDir, err := ioutil.TempDir("", "fabric-mempool_")
-	if err != nil {
-		panic(err)
+	rootDir := os.Getenv("MEMPOOL_DATA")
+	var err error
+	if rootDir == "" {
+		rootDir, err = ioutil.TempDir("", "fabric-mempool_")
+		if err != nil {
+			panic(err)
+		}
+		defer func() {
+			_ = os.RemoveAll(rootDir)
+		}()
 	}
-	defer func() {
-		_ = os.RemoveAll(rootDir)
-	}()
 
 	cfg := config.DefaultMempoolConfig()
 	cfg.CacheSize = 1000
